+++ tests/configs/multi_input.yaml
@@ -63,8 +63,3 @@ model:
             task: binary
       visualizers:
         - name: SegmentationVisualizer
-
-trainer:
-  batch_size: 1
-  epochs: 1
-  validation_interval: 1
