--- tests/configs/parking_lot_config.yaml
@@ -1,93 +1,0 @@
-
-model:
-  name: parking_lot_model
-  nodes:
-
-    - name: EfficientRep
-
-    - name: RepPANNeck
-
-    - name: EfficientBBoxHead
-      task_name: vehicle_type
-      losses:
-        - name: AdaptiveDetectionLoss
-      metrics:
-        - name: MeanAveragePrecision
-          is_main_metric: true
-      visualizers:
-        - name: BBoxVisualizer
-
-    - name: EfficientKeypointBBoxHead
-      task_name: motorbike
-      losses:
-        - name: EfficientKeypointBBoxLoss
-      metrics:
-        - name: MeanAveragePrecision
-      visualizers:
-        - name: KeypointVisualizer
-
-    - name: SegmentationHead
-      task_name: color
-      losses:
-        - name: CrossEntropyLoss
-      metrics:
-        - name: JaccardIndex
-      visualizers:
-        - name: SegmentationVisualizer
-
-    - name: BiSeNetHead
-      task_name: brand
-      losses:
-        - name: CrossEntropyLoss
-      metrics:
-        - name: Precision
-      visualizers:
-        - name: SegmentationVisualizer
-
-    - name: BiSeNetHead
-      task_name: vehicle_type
-      losses:
-        - name: CrossEntropyLoss
-      metrics:
-        - name: Accuracy
-      visualizers:
-        - name: SegmentationVisualizer
-
-tracker:
-  project_name: Parking_Lot
-  is_tensorboard: true
-
-loader:
-  train_view: val
-  params:
-    dataset_name: D1ParkingLot
-
-trainer:
-  accelerator: auto
-  devices: auto
-  strategy: auto
-
-  n_sanity_val_steps: 1
-  profiler: null
-  batch_size: 2
-  accumulate_grad_batches: 1
-  epochs: 200
-  n_workers: 8
-  validation_interval: 10
-  n_log_images: 8
-  skip_last_batch: true
-  log_sub_losses: true
-  save_top_k: 3
-
-  preprocessing:
-    train_image_size: [128, 160]
-    keep_aspect_ratio: false
-    normalize:
-      active: true
-    augmentations:
-      - name: Defocus
-        params:
-          p: 0.1
-      - name: Sharpen
-        params:
-          p: 0.1
